% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ach_marco_polo.R
\name{ach_marco_polo}
\alias{ach_marco_polo}
\title{ach_marco_polo}
\usage{
ach_marco_polo(data)
}
\arguments{
\item{data}{A dataset containing match data, including `event_year`, `fighter_id`, and `event_country`.}
}
\value{
A data frame with fighters who traveled to the most countries in a given year.
}
\description{
Awarded to the fighter who traveled to the most countries (at least 2) during a specific year.
If multiple fighters share the most countries visited, they all receive the achievement.
}
\keyword{internal}
