% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ach_pilgrim.R
\name{ach_pilgrim}
\alias{ach_pilgrim}
\title{The Pilgrim Achievement}
\usage{
ach_pilgrim(data)
}
\arguments{
\item{data}{A data frame with at least the following columns:
\describe{
  \item{fighter_id}{Identifier of the fighter.}
  \item{event_year}{Year of the event.}
  \item{event_country}{Country where the event was held.}
}}
}
\value{
A data frame with columns:
\describe{
  \item{fighter_id}{ID of the fighter.}
  \item{tier_id}{Numeric tier ID.}
  \item{achieved}{Logical, TRUE if achieved.}
  \item{percentile}{Proportion of fighters who achieved this tier.}
  \item{achievement_tier}{Tier name (Bronze, Silver, Gold, Epic).}
  \item{achievement_name}{Name of the achievement.}
  \item{achievement_description}{Description of the achievement.}
  \item{achievement_icon}{Filename of the achievement icon.}
}
}
\description{
Internal function to award the "The Pilgrim" achievement to fighters who traveled to multiple countries to fight in tournaments in a given year.
- Bronze: Participated in tournaments in 2 countries.
- Silver: Participated in tournaments in 5 countries.
- Gold: Participated in tournaments in 15 countries.
- Epic: Participated in the most countries in the year.
}
\keyword{internal}
