% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ach_reaper.R
\name{ach_reaper}
\alias{ach_reaper}
\title{ach_reaper}
\usage{
ach_reaper(data)
}
\arguments{
\item{data}{A dataset containing match data, including `event_year`, `fighter_id`, and `result`.}
}
\value{
A data frame with fighters who won the most fights globally in a given year.
}
\description{
Awarded to the fighter who won the most fights in the world during a specific year. 
If multiple fighters share the most wins, they all receive the achievement.
}
\keyword{internal}
