% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ach_showing_up.R
\name{ach_showing_up}
\alias{ach_showing_up}
\title{Showing Up Achievement}
\usage{
ach_showing_up(data)
}
\arguments{
\item{data}{A data frame with at least the following columns:
\describe{
  \item{fighter_id}{Identifier of the fighter}
  \item{event_year}{Year of the event}
}}
}
\value{
A data frame with columns:
\describe{
  \item{fighter_id}{ID of the fighter}
  \item{tier_id}{Numeric tier ID (1 for Unique)}
  \item{achieved}{Logical, TRUE if achieved}
  \item{percentile}{Proportion of fighters who achieved this tier in that year}
  \item{achievement_tier}{"Unique"}
  \item{achievement_name}{"Showing Up is Part of the Success! - <Year>"}
  \item{achievement_description}{Dynamic description mentioning the year}
  \item{achievement_icon}{"showing_up_unique.png"}
}
}
\description{
Internal function to award the "Showing Up is Part of the Success!" achievement to fighters who participated in a tournament in a given year.
- Unique (tier_id=1): Awarded to all fighters who participated in at least one tournament in a year.
}
\keyword{internal}
