% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_achievements.R
\name{generate_achievements}
\alias{generate_achievements}
\title{Generate Achievements}
\usage{
generate_achievements(path = "./hema_ratings", export_csv = FALSE)
}
\arguments{
\item{path}{Character. Path to the folder containing HEMA data files. Defaults to `./hema_ratings`.}

\item{export_csv}{Bool. Will export the achievements list as CSV to Path.}
}
\value{
A data frame containing all achievements, with columns:
\describe{
  \item{fighter_id}{ID of the fighter.}
  \item{tier_id}{Numeric tier ID (1-4).}
  \item{achieved}{Logical, TRUE if achieved.}
  \item{percentile}{Proportion of fighters who achieved this tier.}
  \item{achievement_tier}{Character. Tier name ("Bronze", "Silver", etc.).}
  \item{achievement_name}{Character. Name of the achievement.}
  \item{achievement_description}{Character. Description of the achievement.}
  \item{achievement_icon}{Character. Icon file name for the achievement.}
}
}
\description{
External function to generate achievements for HEMA ratings. This function
retrieves data from the specified path or uses package data as a fallback.
It computes achievements by invoking all functions prefixed with `ach_` in
the namespace of the `HEMAscRaper` package.
}
\examples{
generate_achievements("./hema_ratings")
}
