% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_new_model.R
\name{generate_new_model}
\alias{generate_new_model}
\title{Generate a New Data Model and Save Processed Data as CSV Files}
\usage{
generate_new_model(data_list, path = "./hema_ratings")
}
\arguments{
\item{data_list}{A named list of data frames. Expected elements include:
\itemize{
  \item \code{hema_countries}
  \item \code{hema_events}
  \item \code{hema_tournaments}
  \item \code{hema_rankings}
  \item \code{hema_match_results}
  \item \code{hema_fighters}
  \item \code{hema_clubs}
  \item \code{hema_achievements}
}}

\item{path}{- the path wehre files will be imported defaults to ./hema_ratings}
}
\value{
A list of processed data frames.
}
\description{
This function processes a list of raw HEMA data frames into a new data model.
It transforms and joins tables for countries, events, tournaments, ratings, results,
fighters, clubs, and achievements; validates the data model using the \code{dm} package;
and then saves the processed tables as CSV files.
}
