% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_travel_data.R
\name{generate_travel_data}
\alias{generate_travel_data}
\title{Generate Travel Data}
\usage{
generate_travel_data(
  year = NULL,
  tournament_weapon = NULL,
  hema_clubs = HEMAscRaper::hema_clubs,
  hema_events = HEMAscRaper::hema_events,
  hema_match_results = HEMAscRaper::hema_match_results
)
}
\arguments{
\item{year}{A numeric value, vector of numeric values, or NULL. Specifies the event years to filter.}

\item{tournament_weapon}{A character vector or NULL. Specifies the tournament weapons to filter.}

\item{hema_clubs}{A dataset containing HEMA clubs. Default: HEMAscRaper::hema_clubs.}

\item{hema_events}{A dataset containing HEMA events. Default: HEMAscRaper::hema_events.}

\item{hema_match_results}{A dataset containing HEMA match results. Default: HEMAscRaper::hema_match_results.}
}
\value{
A list containing two data frames: `nodes` and `edges`.
}
\description{
This is an internal function to generate travel data from HEMA-related datasets.
}
\keyword{internal}
