% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_events.R
\name{get_events}
\alias{get_events}
\title{Download and Extract HEMA Events Data}
\usage{
get_events(raw = FALSE)
}
\arguments{
\item{raw}{Logical. If `TRUE`, returns the raw, unprocessed data as scraped from the website. 
If `FALSE` (default), the data is normalized for consistency (e.g., names, countries, and cities).}
}
\value{
A tibble containing events' details with the following columns:
\describe{
  \item{\code{event_id}}{Unique identifier of the event. Integer.}
  \item{\code{event_brand}}{Name of the event brand. Character.}
  \item{\code{event_name}}{Name of the event. Character.}
  \item{\code{event_year}}{Year in which the event took place. Integer.}
  \item{\code{event_date}}{Date of the event in "YYYY-MM-DD" format. Date.}
  \item{\code{event_country}}{Country where the event was held. Character.}
  \item{\code{event_city}}{City where the event took place, if available. Character.}
  \item{\code{event_url}}{URL of the event's page on the HEMA Ratings website. Character.}
}
}
\description{
This function downloads data from the *HEMA Ratings* website and extracts detailed information
about historical European martial arts events, including their dates, locations, and corresponding URLs.
}
\details{
This function scrapes the HEMA Ratings events page and compiles a list of all HEMA events, including
the event name, date, country, and other relevant details. When the `raw` parameter is set to `FALSE`,
the data is normalized to ensure consistency in names, countries, and cities. 

If city information is missing for an event, the function defaults it to `NA`.

The function assumes the structure of the HEMA Ratings webpage remains consistent. If the webpage layout changes,
this function may need adjustments.
}
\examples{
\dontrun{
# Download HEMA events data and display the first few rows
events_data <- get_events()
head(events_data)

# Retrieve raw, unprocessed events data
raw_events_data <- get_events(raw = TRUE)
}

}
