% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hema_countries}
\alias{hema_countries}
\title{HEMA Countries Dataset}
\format{
A data frame with the following columns:
\describe{
  \item{country_id}{Integer. Unique identifier for the country.}
  \item{name}{Character. Name of the country.}
  \item{region}{Character. Region of the country.}
  \item{sub_region}{Character. Sub-region of the country.}
  \item{population}{Int Active Longsword fencers in community}
  \item{community}{Factor Longsword community id}
  \item{community_label}{Character Community Label}
  \item{country_code}{Character Alpha-2 ISO code}
  
}
}
\usage{
hema_countries
}
\description{
This dataset contains a list of countries associated with HEMA activities.
Note: the community were calculated using the generate community graph for Longsword with a travel threshold of 5 travels
}
\examples{
data(hema_countries)
head(hema_countries)
}
\keyword{datasets}
