% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hema_events}
\alias{hema_events}
\title{HEMA Events Dataset}
\format{
A data frame with the following columns:
\describe{
  \item{event_id}{Integer. Unique identifier for each event.}
  \item{event_name}{Character. Name of the event.}
  \item{event_year}{Integer. Year of the event.}
  \item{event_date}{Date. Date of the event.}
  \item{event_city}{Character. City where the event took place.}
  \item{event_country}{Character. Country of the event.}
  \item{event_brand}{Character. Brand or organizer of the event.}
  \item{event_url}{Character. Website URL of the event.}
}
}
\usage{
hema_events
}
\description{
This dataset contains details about HEMA-related events, including tournaments and other activities.
}
\examples{
data(hema_events)
head(hema_events)
}
\keyword{datasets}
