% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hema_fighters}
\alias{hema_fighters}
\title{HEMA Fighters Dataset}
\format{
A data frame with the following columns:
\describe{
  \item{fighter_id}{Integer. Unique identifier for each fighter.}
  \item{fighter_name}{Character. Name of the fighter.}
  \item{fighter_nationality}{Character. Nationality of the fighter.}
  \item{fighter_club_id}{Integer. Club ID associated with the fighter.}
  \item{fighter_club_name}{Character. Club name associated with the fighter.}
  \item{fighter_url}{Character. Profile URL of the fighter.}
}
}
\usage{
hema_fighters
}
\description{
This dataset contains information about individual HEMA fighters, including their affiliations and metadata.
}
\examples{
data(hema_fighters)
head(hema_fighters)
}
\keyword{datasets}
