% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hema_fights}
\alias{hema_fights}
\title{HEMA Fights Dataset}
\format{
A data frame with the following columns:
\describe{
  \item{match_id}{Integer. Unique identifier for each match.}
  \item{event_id}{Integer. Identifier for the event in which the match took place.}
  \item{event_name}{Character. Name of the event.}
  \item{fighter_1}{Character. Name of the first fighter.}
  \item{fighter_1_result}{Character. Result of the fight for the first fighter (e.g., Win, Loss).}
  \item{fighter_2}{Character. Name of the second fighter.}
  \item{fighter_2_result}{Character. Result of the fight for the second fighter.}
  \item{fighter_id}{Integer. ID of the fighter.}
  \item{opponent_id}{Integer. ID of the opponent.}
  \item{stage}{Character. Stage of the tournament.}
  \item{tournament_category}{Character. Category of the tournament.}
  \item{tournament_id}{Integer. ID of the tournament.}
  \item{tournament_name}{Character. Name of the tournament.}
  \item{tournament_note}{Character. Notes about the tournament.}
  \item{tournament_weapon}{Character. Weapon category used in the tournament.}
}
}
\usage{
hema_fights
}
\description{
This dataset contains information about individual HEMA fights, including the fighters involved and outcomes.
}
\examples{
data(hema_fights)
head(hema_fights)
}
\keyword{datasets}
