% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_names.R
\name{normalize_names}
\alias{normalize_names}
\title{Normalize Names}
\usage{
normalize_names(name)
}
\arguments{
\item{name}{A character vector of names to be cleaned and normalized.}
}
\value{
A character vector of cleaned and normalized names.
}
\description{
This function normalizes names by:
- Removing non-ASCII characters.
- Converting text to lowercase.
- Converting the text to title case.
- Trimming and squishing extra spaces.
}
\examples{
 \dontrun{
# Example usage:
names <- c(
  "JOHN DOE  ", 
  "àLEX smith", 
  "MARY-jane"
)
normalize_names(names)
# [1] "John Doe"   "Alex Smith" "Mary-Jane"
}

}
\keyword{internal}
