% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_fights}
\alias{raw_fights}
\title{HEMA Fights Dataset}
\format{
A data frame with the following columns:
\describe{
  \item{\code{event_id}}{Unique identifier of the event. Integer.}
  \item{\code{event_name}}{Name of the event. Character.}
  \item{\code{tournament_name}}{Name of the specific tournament or discipline within the event. Character.}
  \item{\code{tournament_category}}{Category of the tournament (e.g., Mixed, Men's, Women's, Underrepresented Genders). Character.}
  \item{\code{tournament_note}}{Other notes. Character.}
  \item{\code{tournament_weapon}}{Weapon used in the tournament (e.g., Longsword, Rapier). Character.}
  \item{\code{fighter_id}}{Unique identifier of the first fighter in a match. Integer.}
  \item{\code{opponent_id}}{Unique identifier of the opponent fighter in a match. Integer.}
  \item{\code{stage}}{Name of the stage. Character.}
  \item{\code{fighter_1}}{Name of the fighter Character.}
  \item{\code{fighter_2}}{Name of the fighter Character.}
  \item{\code{fighter_1_result}}{Score of the first fighter. Character.}
  \item{\code{fighter_2_result}}{Score of the opponent fighter. Character.}
}
}
\source{
[HEMA Ratings Tournament Page](https://hemaratings.com/tournaments/)
}
\usage{
raw_fights
}
\description{
This dataset contains information about individual fights in Historical European Martial Arts (HEMA) tournaments, including fighters' scores, match outcomes, and tournament details.
}
\details{
The dataset provides detailed match-level statistics for HEMA tournaments. It can be used to analyze match outcomes, fighter performance,
and trends across different weapon categories and tournament types.
}
\examples{
data(raw_fights)
summary(raw_fights)

}
\keyword{datasets}
