% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{HEMA Schievement Test Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10000 rows and 43 columns.
}
\usage{
test_data
}
\description{
This dataset contains aggregated information about HEMA tournaments, including participant counts and match statistics.
}
\examples{
data(test_data)
head(test_data)
}
\keyword{datasets}
