% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ach_consistency.R
\name{ach_consistency_is_key}
\alias{ach_consistency_is_key}
\title{ach_consistency_is_key}
\usage{
ach_consistency_is_key(data)
}
\arguments{
\item{data}{A dataset containing match data, including `event_year`, `fighter_id`, and `event_id`.}
}
\value{
A data frame with fighters who participated in the most events in a given year.
}
\description{
Awarded to the fighter who participated in the most events during a specific year.
If multiple fighters share the highest number of events participated in, they all receive the achievement.
}
\keyword{internal}
