% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ach_like_cozy.R
\name{ach_like_cozy}
\alias{ach_like_cozy}
\title{I like it Cozy Achievement}
\usage{
ach_like_cozy(data)
}
\arguments{
\item{data}{A data frame with at least the following columns:
\describe{
  \item{fighter_id}{Identifier of the fighter}
  \item{club_country}{Country of the fighter's club}
  \item{event_country}{Country where the event took place}
  \item{match_id}{Identifier of the match}
}}
}
\value{
A data frame with columns:
\describe{
  \item{fighter_id}{ID of the fighter}
  \item{tier_id}{Numeric tier ID (1)}
  \item{achieved}{Logical, always TRUE for achievers}
  \item{percentile}{Proportion of fighters who achieved this}
  \item{achievement_tier}{"Unique"}
  \item{achievement_name}{"I like it Cozy"}
  \item{achievement_description}{Includes the total number of home matches}
  \item{achievement_icon}{"i_like_it_cozy_unique.png"}
}
}
\description{
Internal function to award the "I like it Cozy" achievement to fighters who have never fought outside their home country
and have fought more than 100 matches.
}
\keyword{internal}
