% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ach_make_friends.R
\name{ach_make_friends_weapon}
\alias{ach_make_friends_weapon}
\title{It Is Good to Make Friends Achievement (Per Weapon)}
\usage{
ach_make_friends_weapon(data)
}
\arguments{
\item{data}{A data frame with at least the following columns:
\describe{
  \item{fighter_id}{Identifier of the fighter.}
  \item{event_year}{Year of the event.}
  \item{opponent_id}{Identifier of the opponent.}
  \item{tournament_weapon}{The weapon category of the tournament.}
}}
}
\value{
A data frame with columns:
\describe{
  \item{fighter_id}{ID of the fighter.}
  \item{tier_id}{Numeric tier ID.}
  \item{achieved}{Logical, TRUE if achieved.}
  \item{percentile}{Proportion of fighters who achieved this tier, relative to all fighters.}
  \item{achievement_tier}{Tier name (Unique).}
  \item{achievement_name}{Name of the achievement.}
  \item{achievement_description}{Description of the achievement.}
  \item{achievement_icon}{Filename of the achievement icon.}
}
}
\description{
Internal function to award the "It Is Good to Make Friends" achievement to fighters who faced the most unique opponents in a specific weapon category and year. Ties are allowed.
}
\keyword{internal}
