% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ach_rank.R
\name{ach_rank_longsword}
\alias{ach_rank_longsword}
\title{ach_rank_longsword}
\usage{
ach_rank_longsword(data)
}
\arguments{
\item{data}{A dataset containing match data, including `tournament_weapon`, `event_year`, and `result`.}
}
\value{
A data frame of fighters, their ranks, and corresponding achievements for each weapon and year.
}
\description{
Find the top fighters for each tournament weapon and year! Fighters are ranked based on their Bayesian win percentage, and achievements are awarded:
- Epic: Top 1% of fighters.
- Gold: Top 5% of fighters.
- Silver: Top 15% of fighters.
- Bronze: Top 30% of fighters.
}
\keyword{internal}
