% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HEMAscrapers.R
\name{get_events}
\alias{get_events}
\title{Download and Extract HEMA Events Data}
\usage{
get_events()
}
\value{
A tibble containing events' details with the following columns:
\describe{
  \item{\code{event_id}}{Unique identifier of the event. Integer.}
  \item{\code{event_name}}{Name of the event. Character.}
  \item{\code{event_year}}{Year in which the event took place. Integer.}
  \item{\code{event_date}}{Date of the event in "YYYY-MM-DD" format. Date.}
  \item{\code{event_country}}{Country where the event was held. Character.}
  \item{\code{event_city}}{City where the event took place, if available. Character.}
  \item{\code{event_url}}{URL of the event's page on the HEMA Ratings website. Character.}
}
}
\description{
This function downloads data from the *HEMA Ratings* website and extracts detailed information
about historical European martial arts events, including their dates, locations, and corresponding URLs.
}
\details{
This function scrapes the HEMA Ratings events page and compiles a list of all HEMA events, including
the event name, date, country, and other relevant details. It uses the `rvest` and `dplyr` packages
to parse the HTML structure and organize the extracted information into a tibble.

The function assumes the structure of the HEMA Ratings webpage remains consistent. If the webpage layout changes,
this function may need adjustments.
}
\note{
The function handles cases where city information might be missing and defaults to `NA` when a city is not listed.
}
\examples{
\dontrun{
# Download HEMA events data and display the first few rows
events_data <- get_events()
head(events_data)
}

}
