% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hema_fighters}
\alias{hema_fighters}
\title{HEMA Fighters Dataset}
\format{
A data frame with the following columns:
\describe{
  \item{\code{fighter_id}}{Unique identifier of the fighter. Integer.}
  \item{\code{fighter_name}}{Name of the fighter. Character.}
  \item{\code{fighter_nationality}}{Nationality of the fighter (based on country flag displayed). Character.}
  \item{\code{fighter_club_id}}{Unique identifier of the club, if available. Integer.}
  \item{\code{fighter_club_name}}{Name of the fighter's club, if available. Character.}
  \item{\code{fighter_url}}{URL of the fighter's profile on the HEMA Ratings website. Character.}
}
}
\source{
[HEMA Ratings Fighters Page](https://hemaratings.com/fighters/)
}
\usage{
hema_fighters
}
\description{
This dataset contains information about registered Historical European Martial Arts (HEMA) fighters, including their names, nationalities, and affiliated clubs.
}
\details{
This dataset provides a comprehensive overview of the fighters registered on the HEMA Ratings website.
It can be used for demographic analysis of HEMA practitioners, visualizing the distribution of fighters, and linking fighters to their clubs.
}
\examples{
data(hema_fighters)
head(hema_fighters)

}
\keyword{datasets}
